requirements.txt
python-telegram-bot
python-dotenv
deepeval
openai
